# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pawzee API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type HazardCategory =
  (typeof HazardCategory)[keyof typeof HazardCategory];

export const HazardCategory = {
  broken_glass: "broken_glass",
  poison_bait: "poison_bait",
  aggressive_dog: "aggressive_dog",
  construction: "construction",
  spray_activity: "spray_activity",
  ticks_fleas: "ticks_fleas",
  stray_animal: "stray_animal",
  flooding: "flooding",
  ant_nest: "ant_nest",
  waste: "waste",
  other: "other",
} as const;

export interface HazardItem {
  id: string;
  category: HazardCategory;
  lat: number;
  lng: number;
  /** @nullable */
  photoUrl?: string | null;
  reportedBy: string;
  /** @nullable */
  reportedByName?: string | null;
  reportedAt: string;
  expiresAt: string;
  confirmationCount: number;
}

export interface HazardListResponse {
  hazards: HazardItem[];
}

export interface CreateHazardRequest {
  category: HazardCategory;
  lat: number;
  lng: number;
  /** @nullable */
  photoUrl?: string | null;
}

export interface ConfirmHazardRequest {
  lat: number;
  lng: number;
}

export type HazardSummaryResponseBreakdown = { [key: string]: number };

export interface HazardSummaryResponse {
  hazardsToday: number;
  activeHazards: number;
  breakdown: HazardSummaryResponseBreakdown;
}

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListHazardsParams = {
  lat: number;
  lng: number;
  /**
   * Radius in meters (default 5000)
   */
  radius?: number;
};

export type GetHazardSummaryParams = {
  lat: number;
  lng: number;
  radius?: number;
};