# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pawzee API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: hazards
    description: Hazard reporting and viewing
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect on success or failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
  /hazards:
    get:
      tags: [hazards]
      operationId: listHazards
      summary: List hazards within a radius
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
        - name: radius
          in: query
          required: false
          description: Radius in meters (default 5000)
          schema:
            type: number
      responses:
        '200':
          description: List of hazards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HazardListResponse'
    post:
      tags: [hazards]
      operationId: createHazard
      summary: Report a new hazard
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHazardRequest'
      responses:
        '201':
          description: Hazard created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HazardItem'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /hazards/{id}/confirm:
    post:
      tags: [hazards]
      operationId: confirmHazard
      summary: Confirm a hazard (must be within 10m)
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmHazardRequest'
      responses:
        '200':
          description: Hazard confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HazardItem'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '400':
          description: Too far from hazard or already confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /hazards/summary:
    get:
      tags: [hazards]
      operationId: getHazardSummary
      summary: Get hazard summary for an area
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
        - name: radius
          in: query
          required: false
          schema:
            type: number
      responses:
        '200':
          description: Hazard summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HazardSummaryResponse'
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    HazardCategory:
      type: string
      enum:
        - broken_glass
        - poison_bait
        - aggressive_dog
        - construction
        - spray_activity
        - ticks_fleas
        - stray_animal
        - flooding
        - ant_nest
        - waste
        - other
    HazardItem:
      type: object
      required: [id, category, lat, lng, reportedBy, reportedAt, expiresAt, confirmationCount]
      properties:
        id:
          type: string
        category:
          $ref: '#/components/schemas/HazardCategory'
        lat:
          type: number
        lng:
          type: number
        photoUrl:
          type: [string, 'null']
        reportedBy:
          type: string
        reportedByName:
          type: [string, 'null']
        reportedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        confirmationCount:
          type: integer
    HazardListResponse:
      type: object
      required: [hazards]
      properties:
        hazards:
          type: array
          items:
            $ref: '#/components/schemas/HazardItem'
    CreateHazardRequest:
      type: object
      required: [category, lat, lng]
      properties:
        category:
          $ref: '#/components/schemas/HazardCategory'
        lat:
          type: number
        lng:
          type: number
        photoUrl:
          type: [string, 'null']
    ConfirmHazardRequest:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number
    HazardSummaryResponse:
      type: object
      required: [hazardsToday, activeHazards, breakdown]
      properties:
        hazardsToday:
          type: integer
        activeHazards:
          type: integer
        breakdown:
          type: object
          additionalProperties:
            type: integer