# lib/api-zod/src/generated/types/confirmHazardRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pawzee API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConfirmHazardRequest {
  lat: number;
  lng: number;
}