# lib/api-zod/src/generated/types/createHazardRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pawzee API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HazardCategory } from "./hazardCategory";

export interface CreateHazardRequest {
  category: HazardCategory;
  lat: number;
  lng: number;
  /** @nullable */
  photoUrl?: string | null;
}