# lib/api-zod/src/generated/types/hazardCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pawzee API specification
 * OpenAPI spec version: 0.1.0
 */

export type HazardCategory =
  (typeof HazardCategory)[keyof typeof HazardCategory];

export const HazardCategory = {
  broken_glass: "broken_glass",
  poison_bait: "poison_bait",
  aggressive_dog: "aggressive_dog",
  construction: "construction",
  spray_activity: "spray_activity",
  ticks_fleas: "ticks_fleas",
  stray_animal: "stray_animal",
  flooding: "flooding",
  ant_nest: "ant_nest",
  waste: "waste",
  other: "other",
} as const;