# lib/api-zod/src/generated/types/hazardItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pawzee API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HazardCategory } from "./hazardCategory";

export interface HazardItem {
  id: string;
  category: HazardCategory;
  lat: number;
  lng: number;
  /** @nullable */
  photoUrl?: string | null;
  reportedBy: string;
  /** @nullable */
  reportedByName?: string | null;
  reportedAt: Date;
  expiresAt: Date;
  confirmationCount: number;
}