# lib/api-zod/src/generated/types/hazardListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pawzee API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HazardItem } from "./hazardItem";

export interface HazardListResponse {
  hazards: HazardItem[];
}