# lib/api-zod/src/generated/types/hazardSummaryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pawzee API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HazardSummaryResponseBreakdown } from "./hazardSummaryResponseBreakdown";

export interface HazardSummaryResponse {
  hazardsToday: number;
  activeHazards: number;
  breakdown: HazardSummaryResponseBreakdown;
}